requirements.txt

email-validator>=2.2.0
flask-session>=0.8.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
werkzeug>=3.1.3
pytz
openai
openai
sendgrid
twilio
twilio
flask-dance
flask-login
oauthlib
pyjwt
requests
sendgrid
twilio
pillow
reportlab
pillow
